v2.0.0 Release
Added two new configuration options, using /gamerule. These options affect the right click Ender Pearl ability cooldown, as well as the durability cost when using this ability. Additionally, the sword textures have been redone, the crafting recipe for the Endersword Hilt has been changed to make more sense, and the durability costs have been slightly tweaked on the swords.

The default ability cooldown percentage is 100. Lower this value to speed up the cooldown time, and raise it to make the cooldown longer.

The default durability cost multiplier is 1. Lower this value to 0 to disable durability cost, and raise it up to 2 to double the durability cost.